hiding keyboard moved to fragment
