+++ app/src/main/java/intive/ideabox/activity/MainActivity.java
@@ -1,7 +1,7 @@
 package intive.ideabox.activity;
 
-import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
+import android.support.v7.app.AppCompatActivity;
 
 import intive.ideabox.R;
 import intive.ideabox.fragment.IdeaListFragment;
@@ -12,7 +12,7 @@ public class MainActivity extends AppCompatActivity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
-
         getSupportFragmentManager().beginTransaction().replace(R.id.fragmentContainer, IdeaListFragment.newInstance(false)).commit();
     }
+
 }
