+++ app/src/main/res/layout/fragment_add_idea.xml
@@ -20,9 +20,10 @@
         <RelativeLayout
             android:layout_width="match_parent"
             android:layout_height="match_parent"
+            android:focusableInTouchMode="true"
             tools:context="intive.ideabox.activity.MainActivity">
             <ScrollView
-                android:id="@+id/scroll"
+                android:id="@+id/backgroundLayout"
                 android:layout_width="match_parent"
                 android:layout_height="match_parent">
                 <ImageView
